Fixed a large number of WIN64 type conversion warnings.
Changed FS_Write to do multiple write calls if the size exceeds 1GB, this is to accommodate WIN64 where write() takes a UINT rather than size_t.  We have no use of such large writes at this time, so this is a mostly theoretical feature.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12081 d7cf8633-e32d-0410-b094-e92efae38249

+++ cl_parse.c
@@ -1377,7 +1377,7 @@ static void CL_StopDownload(int size, int crc)
 			{
 				Con_Printf("Inflated download: new size: %u (%g%%)\n", (unsigned)inflated_size, 100.0 - 100.0*(cls.qw_downloadmemorycursize / (float)inflated_size));
 				cls.qw_downloadmemory = out;
-				cls.qw_downloadmemorycursize = inflated_size;
+				cls.qw_downloadmemorycursize = (int)inflated_size;
 			}
 			else
 			{
